Changed health values to support floats

+++ Assets/UI/Scripts/HUD.cs
@@ -39,7 +39,7 @@ public class HUD : MonoBehaviour
         timerDisplay.text = time.ToString("mm':'ss");
     }
 
-    public void UpdateHealthBar(int newHP)
+    public void UpdateHealthBar(float newHP)
     {
         healthBarDisplay.UpdateHealthBar(newHP);
     }
